move json editor in separate module
